Problem Sovling 1463 (number of challenges : 19)
There are three operations that can be used on integer X as follows.

If X is divisible by 3, divide by 3.
If X is divisible by 2, divide by 2.
Subtract 1.
Given an integer N, we try to make it 1 by appropriately using the three operations above. Output the minimum number of times the operation is used.